Fix build: add GlobalUsings.cs for WPF temp project, CS0121/CS0173 compile errors, System.IO.Ports package
- Add GlobalUsings.cs with explicit global using directives so the WPF XAML
  compiler's auto-generated *_wpftmp.csproj (which does not inherit
  <ImplicitUsings>) can resolve System, IO, Linq, Threading, etc.
- Add System.IO.Ports 8.0.0 NuGet package (SerialPort not in default refs)
- Fix CS0173 in MainViewModel.ApplyUpdate: change var to object? so both
  null conditional branches have an unambiguous type
- Fix CS0121 in EasterEggControl: Math.Max(g, (byte)0) removes ambiguity
  between Math.Max(byte,byte) and Math.Max(int,int) overloads

+++ src/PortPane/Views/EasterEggControl.cs
@@ -134,7 +134,7 @@ public sealed class EasterEggControl : Canvas
                 var tb = new TextBlock
                 {
                     Text       = chars[_rng.Next(chars.Length)].ToString(),
-                    Foreground = new SolidColorBrush(Color.FromRgb(0, (byte)Math.Max(g, 0), 0)),
+                    Foreground = new SolidColorBrush(Color.FromRgb(0, (byte)Math.Max(g, (byte)0), 0)),
                     FontFamily = new FontFamily("Consolas"),
                     FontSize   = 11
                 };
